// packages/twenty-server/src/modules/workflow/workflow-executor/workflow-actions/create-calendar-event/types/workflow-create-calendar-event-action-input.type.ts
export type WorkflowCreateCalendarEventActionInput = {
  connectedAccountId: string;
  title: string;
  description?: string;
  location?: string;
  startsAt: string;
  endsAt: string;
  isFullDay: boolean;
  timeZone?: string;
  attendees?: string;
  sendInvitations: boolean;
  addConferencing: boolean;
};